plugins/template-work/skills/extraction-tool/requirements.txt
PyYAML>=6.0
glob2>=0.7
shutilwhich>=1.1.0